Refactor vacation store out of overtime and into vacation.ts
