+++ packages/frontend/src/store/overtime.ts
@@ -16,24 +16,6 @@ export interface OvertimeStateModel {
   compansatedHours: number;
   totalFlexedHours: number;
   totalPayoutHours: number;
-  vacationEntries: EntriesModel[],
-  totalVacationHoursUsed: number,
-  totalVacationDaysUsed: number
-}
-
-interface VacationTimeModel {
-  totalHoursUsed: number;
-  totalDaysUsed: number;
-  entries: EntriesModel[];
-}
-
-interface EntriesModel {
-  user: number;
-  userEmail: string;
-  id: number;
-  date: Date;
-  value: number;
-  taskId: number;
 }
 
 export interface CompansatedTransactions {
@@ -101,9 +83,6 @@ const initState: OvertimeStateModel = {
   compansatedHours: 0,
   totalFlexedHours: 0,
   totalPayoutHours: 0,
-  vacationEntries: [],
-  totalVacationDaysUsed: 0,
-  totalVacationHoursUsed: 0,
 };
 
 const state: OvertimeState = {
@@ -201,34 +180,9 @@ const getters = {
   getAvailableCompensated: (state: State) => {
     return state.overtimeState.compansatedHours;
   },
-  getUsedVacationEntries: (state: State) => {
-    return state.overtimeState.vacationEntries;
-  },
-  getUsedVacationHours: (state: State) => {
-    return state.overtimeState.totalVacationHoursUsed;
-  },
-  getUsedVacationDays: (state: State) => {
-    return state.overtimeState.totalVacationDaysUsed;
-  },
 };
 
 const actions = {
-  FETCH_USED_VACATION: async ({ commit }: ActionContext<State, State>, parameters: { year: number }) => {
-    try {
-        let url = new URL(
-          config.API_HOST + `/api/user/UsedVacation?year=${parameters.year}`
-        ).toString();
-        const response = await adAuthenticatedFetch(url);
-        const data = await response.json();
-              if (response.status === 404) {
-        commit("SET_USER_NOT_FOUND");
-        throw response.statusText;
-      }
-        commit("SET_USEDVACATION", data);
-      } catch (e) {
-        console.error(e);
-      }
-    },
   FETCH_AVAILABLE_HOURS: async ({ commit }: ActionContext<State, State>) => {
     try {
       let url = new URL(
@@ -344,11 +298,6 @@ const mutations = {
       transactions.availableHoursBeforeCompensation;
     state.overtimeState.compansatedHours =
       transactions.availableHoursAfterCompensation;
-  },
-    SET_USEDVACATION(state: State, transactions: VacationTimeModel) {
-    state.overtimeState.vacationEntries = transactions.entries;
-    state.overtimeState.totalVacationHoursUsed = transactions.totalHoursUsed;
-    state.overtimeState.totalVacationDaysUsed = transactions.totalDaysUsed;
   },
 };
